fix: move `Position` to separate module
